added functionality to view other profiles and to be able to click profiles from reviews
