+++ src/Components/App.js
@@ -3,15 +3,15 @@ import {
   BrowserRouter,
   Route,
   Routes,
-  Link
 } from "react-router-dom";
 import LandingPage from "./LandingPage"
 import Login from "./Login"
 import Feed from "./Feed"
 import Error from "./Error"
 import Movie from "./Movie"
-import Profile from "./Profile"
+import CreateProfile from "./CreateProfile"
 import ReviewCreator from "./ReviewCreator"
+import Profile from "./Profile";
 import "./App.css";
 import { AuthContextProvider } from "../Context/AuthContext";
 
@@ -27,7 +27,8 @@ function App() {
               <Route path="/login" element= {<Login />}/>
               <Route path="/feed" element= {<Feed />}/>
               <Route path="*" element= {<Error/>}/>
-              <Route path="/profile" element= {<Profile/>}/>
+              <Route path="/create-profile" element= {<CreateProfile/>}/>
+              <Route path="/profile/:userId" element= {<Profile/>}/>
               <Route path="/movie/:movieId/:movietitle" element= {<Movie/>}/>
               <Route path="/create-review" element = {<ReviewCreator/>}/>
             </Routes>
