+++ src/Components/CreateProfile.css
@@ -1,4 +1,4 @@
-.profile-picture {
+.profile-pic {
   height: 30vw;
   width: 30vw;
   object-fit: cover;
