+++ src/firebase.js
